Replaced VariableNameMinLengthRule with new version using ConfigurableRule
