+++ Source/SwiftLintFramework/Rules/VariableNameMinLengthRule.swift
@@ -1,27 +1,40 @@
 //
-//  VariableNameMinLengthRule.swift
+//  VariableNameMinLengthNewRule.swift
 //  SwiftLint
 //
-//  Created by Mickaël Morier on 04/11/2015.
+//  Created by Scott Hoyt on 01/02/2016.
 //  Copyright © 2015 Realm. All rights reserved.
 //
 
 import SourceKittenFramework
 import SwiftXPC
 
-public struct VariableNameMinLengthRule: ASTRule, ParameterizedRule, ConfigurableRule {
-    public init() {
-        self.init(parameters: [
-            RuleParameter(severity: .Warning, value: 3),
-            RuleParameter(severity: .Error, value: 2)
-        ])
+public struct VariableNameMinLengthRule: ASTRule, ConfigurableRule {
+
+    public init() { }
+
+    public init?(config: AnyObject) {
+        self.init()
+        if let warningNumber = config as? Int {
+            warning = RuleParameter(severity: .Warning, value: warningNumber)
+        } else if let config = config as? [String: AnyObject] {
+            if let warningNumber = config["warning"] as? Int {
+                warning = RuleParameter(severity: .Warning, value: warningNumber)
+            }
+            if let errorNumber = config["error"] as? Int {
+                error = RuleParameter(severity: .Error, value: errorNumber)
+            }
+            if let excluded = config["excluded"] as? [String] {
+                self.excluded = excluded
+            }
+        } else {
+            return nil
+        }
     }
 
-    public init(parameters: [RuleParameter<Int>]) {
-        self.parameters = parameters
-    }
-
-    public let parameters: [RuleParameter<Int>]
+    public var excluded = [String]()
+    private var warning = RuleParameter(severity: .Warning, value: 3)
+    private var error = RuleParameter(severity: .Error, value: 2)
 
     public static let description = RuleDescription(
         identifier: "variable_name_min_length",
@@ -42,15 +55,28 @@ public struct VariableNameMinLengthRule: ASTRule, ParameterizedRule, Configurabl
     public func validateFile(file: File, kind: SwiftDeclarationKind,
                              dictionary: XPCDictionary) -> [StyleViolation] {
         return file.validateVariableName(dictionary, kind: kind).map { name, offset in
-            let charCount = name.characters.count
-            for parameter in parameters.reverse() where charCount < parameter.value {
-                return [StyleViolation(ruleDescription: self.dynamicType.description,
-                    severity: parameter.severity,
-                    location: Location(file: file, byteOffset: offset),
-                    reason: "Variable name should be \(parameter.value) characters " +
-                            "or more: currently \(charCount) characters")]
+            if !excluded.contains(name) {
+                let charCount = name.characters.count
+                for parameter in [error, warning] where charCount < parameter.value {
+                    return [StyleViolation(ruleDescription: self.dynamicType.description,
+                        severity: parameter.severity,
+                        location: Location(file: file, byteOffset: offset),
+                        reason: "Variable name should be \(parameter.value) characters " +
+                        "or more: currently \(charCount) characters")]
+                }
             }
             return []
         } ?? []
+    }
+
+    public func isEqualTo(rule: ConfigurableRule) -> Bool {
+        if let rule = rule as? VariableNameMinLengthRule {
+            // Need to use alternate method to compare excluded due to apparent bug in
+            // the way that SwiftXPC compares [String]
+            return self.error == rule.error &&
+                   self.warning == rule.warning &&
+                   zip(self.excluded, rule.excluded).reduce(true) { $0 && ($1.0 == $1.1) }
+        }
+        return false
     }
 }
